divided the code in separate pages for better functionality
